Move `add_extraction_pipeline` decorator to util module
Single function modules is not very pythonic. Better to keep it in the
util module
